moved trivial ui components
