+++ Decsys/ClientApp/.storybook/config.js
@@ -2,7 +2,7 @@ import React from "react";
 import { configure, addDecorator } from "@storybook/react";
 import { withThemes } from "storybook-styled-components";
 import { Normalize } from "@smooth-ui/core-sc";
-import theme from "../src/themes";
+import theme from "../src/app/themes";
 
 addDecorator(story => (
   <>
